requirements.txt
pyln-client>=0.7.3